fix Pain weapons bug
https://cdn.discordapp.com/attachments/293493549758939136/461236206131478538/PainWeaponBug.gif

because I'm new to coding it took me 2 hours to find this bug. I spent a while looking at the POWERUP_ARMOR and RemoveNinja() code wondering why it didn't work.

this bug is funny to some players so maybe you don't want to fix it. i wont mind because it was a challange for me to find it